+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2101.1.8]
+
+### Added
+* Added client config "Reticle Lines" (default true) to minimap settings, to control reticle crosshairs draw on the minimap
+* Added client config "Show Text above Minimap" (default false) to control if minimap info is shown above or below minimap
+* Waypoints may now be edited from the waypoint manager either by double-clicking, or right-clicking and selecting "Edit" from the context menu
+* Added tr_tr translation (thanks @RuyaSavascisi)
+* Added `#create:table_cloths` to `ftbchunks:interact_whitelist` to allow Create Table Cloths (shop blocks) to be right-clicked anywhere
+
+### Changed
+* Now using FTB Library 2101.1.10 config system
+  * **IMPORTANT NOTE FOR PACK MAKERS** any customized FTB Chunks configuration distributed in `defaultconfigs/ftbchunks/` must now be distributed in `config/` !
+  * `defaultconfigs/ftbchunks/ftbchunks-client.snbt` -> `config/ftbchunks-client.snbt`
+  * `defaultconfigs/ftbchunks/ftbchunks-world.snbt` -> `config/ftbchunks-world.snbt`
+  * Server admins may locally copy `config/ftbchunks-world.snbt` to `world/serverconfig/ftbchunks-world.snbt` if they wish to maintain custom settings which will not be reset by modpack updates
+* Improved entity icons for entities from the Aquaculture mod
+* Updated ru_ru translation (thanks @BazZziliuS)
+
+### Fixed
+* Fixed light technical blocks (`minecraft:light`) rendering black on the map
+* Fixed a piston protection issue which allowed blocks to be pulled out of protected chunks with a sticky piston
+
 ## [2101.1.7]
 
 ### Fixed